feat(syncengine): integrate ResizablePool with adaptive scaling (Phase 4)
Sync engine now detects and resizes SFTP connection pools dynamically:
- Detect ResizablePool interface in NewEngine via type assertion
- Add resizePools() method to resize source/dest pools
- Call resizePools() from MakeScalingDecision() on worker changes
- Call resizePools() from syncAdaptive() during initial setup
- Add 11 comprehensive tests (6 pass, 5 skip pending SFTP infra)

Pool size now follows desiredWorkers count, enabling efficient
resource utilization that adapts to workload (Issue #12 Phase 4).

AI-Used: [claude]

+++ internal/syncengine/sync_test_helpers.go
@@ -15,8 +15,8 @@ func (e *Engine) GetDesiredWorkers() int32 {
 }
 
 // SetDesiredWorkers sets the desired worker count (test helper)
-func (e *Engine) SetDesiredWorkers(count int32) {
-	atomic.StoreInt32(&e.desiredWorkers, count)
+func (e *Engine) SetDesiredWorkers(count int) {
+	atomic.StoreInt32(&e.desiredWorkers, int32(count))
 }
 
 // TestWorker exposes the worker function for testing
